Moves specific list styles from base/defaults into modules/lists
Because the defaults are declared on the `<ul>` and `<ol>` tags, anything qualifying those either need to also be tag-qualified, declared inside this default file, or use `!important`. Those aren't great options. Moving these out into a modules file allows these to be extended more easily and work together on a single element without the ones declared in defaults *always* overpowering the others.