Removes npm-shrinkwrap
After internal discussions, it has been decided that the repo itself should remain un-wrapped, as it were, and shrink-wrapping should be left to the user at the outset of each new project. Right now we're merely recommending and documenting this in the README; in the future it will get rolled up in the build scripts as an option.